feat: ported camera threejs to r3f
